backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai>=1.40.0
python-multipart>=0.0.9
pdf2image>=1.17.0
Pillow>=10.0.0
jinja2>=3.1.0
